Assume rest bug PoCs to be under public domain
They are either not sufficiently large or most likely machine-generated,
as it is unpropable a human would manually write or edit binary formats
randomly, since the corresponding bug report lacks the reasoning
of the bug or the bug-causing conditions.

I could very much make a mistake here though, please let me know
if you are an author of one of these input and want it to be distributed
under different terms.